+++ Sources/Sentry/include/SentryHttpInterceptor+Private.h
@@ -1,0 +1,9 @@
+#include "SentryHttpInterceptor.h"
+
+static NSString *const SENTRY_INTERCEPTED_REQUEST = @"SENTRY_INTERCEPTED_REQUEST";
+
+@interface SentryHttpInterceptor (private)
+
++ (void)configureSessionConfiguration:(NSURLSessionConfiguration *)configuration;
+
+@end